almost finished with officer api

+++ DurableLoans.DomainModel/LoanAmount.cs
@@ -2,7 +2,7 @@
 {
     public class LoanAmount
     {
-        public decimal Amount { get; set; }
+        public int Amount { get; set; }
 
         public string CurrencyType { get; set; }
     }
